[Fix] Refactoring Migrations.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -28,6 +28,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if(app()->isLocal()) {
+            Schema::dropIfExists('users');
+        }
     }
 };
