+++ database/migrations/2023_09_08_081556_create_categories_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('categories');
+        if(app()->isLocal()) {
+            Schema::dropIfExists('categories');
+        }
     }
 };
