+++ database/migrations/2023_09_08_083604_create_article_category_table.php
@@ -35,6 +35,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('article_category');
+        if(app()->isLocal()) {
+            Schema::dropIfExists('article_category');
+        }
     }
 };
