+++ database/migrations/2023_09_11_044344_create_comments_table.php
@@ -45,6 +45,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('comments');
+        if(app()->isLocal()) {
+            Schema::dropIfExists('comments');
+        }
     }
 };
